+++ lib/standard/creates_config_store.rb
@@ -1,62 +1,23 @@
 require "rubocop"
-require "pathname"
+
+require_relative "creates_config_store/assigns_rubocop_yaml"
+require_relative "creates_config_store/sets_target_ruby_version"
+require_relative "creates_config_store/configures_ignored_paths"
 
 module Standard
   class CreatesConfigStore
+    def initialize
+      @assigns_rubocop_yaml = AssignsRubocopYaml.new
+      @sets_target_ruby_version = SetsTargetRubyVersion.new
+      @configures_ignored_paths = ConfiguresIgnoredPaths.new
+    end
+
     def call(standard_config)
-      config_store = RuboCop::ConfigStore.new
-      config_store.options_config = rubocop_yaml_path(standard_config[:ruby_version])
-      options_config = config_store.instance_variable_get("@options_config")
-
-      options_config["AllCops"]["TargetRubyVersion"] = floatify_version(
-        max_rubocop_supported_version(standard_config[:ruby_version])
-      )
-
-      ignored_patterns(standard_config).each do |(path, cops)|
-        cops.each do |cop|
-          options_config[cop] ||= {}
-          options_config[cop]["Exclude"] ||= []
-          options_config[cop]["Exclude"] |= [
-            Pathname.new(standard_config[:config_root]).join(path).to_s,
-          ]
-        end
+      RuboCop::ConfigStore.new.tap do |config_store|
+        options_config = @assigns_rubocop_yaml.call(config_store, standard_config)
+        @sets_target_ruby_version.call(options_config, standard_config)
+        @configures_ignored_paths.call(options_config, standard_config)
       end
-
-      config_store
-    end
-
-    private
-
-    def ignored_patterns(standard_config)
-      standard_config[:ignore]
-    end
-
-    def rubocop_yaml_path(desired_version)
-      file_name = if desired_version < Gem::Version.new("1.9")
-        "ruby-1.8.yml"
-      elsif desired_version < Gem::Version.new("2.0")
-        "ruby-1.9.yml"
-      elsif desired_version < Gem::Version.new("2.3")
-        "ruby-2.2.yml"
-      else
-        "base.yml"
-      end
-
-      Pathname.new(__dir__).join("../../config/#{file_name}")
-    end
-
-    def max_rubocop_supported_version(desired_version)
-      rubocop_supported_version = Gem::Version.new("2.2")
-      if desired_version < rubocop_supported_version
-        rubocop_supported_version
-      else
-        desired_version
-      end
-    end
-
-    def floatify_version(version)
-      major, minor = version.segments
-      "#{major}.#{minor}".to_f # lol
     end
   end
 end
